requirements.txt
absl-py==2.1.0
affine==2.4.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
altair==5.5.0
annotated-types==0.7.0
anyio==4.10.0
appnope==0.1.4
arabic-reshaper==3.0.0
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
arrow==1.3.0
asgiref==3.8.1
asttokens==2.4.1
astunparse==1.6.3
async-lru==2.0.5
attrs==24.2.0
audioread==3.0.1
babel==2.17.0
backcall==0.2.0
backoff==2.2.1
beautifulsoup4==4.13.4
bleach==6.2.0
blinker==1.9.0
blis==1.0.1
branca==0.8.0
cachetools==6.1.0
Cartopy==0.25.0
catalogue==2.0.10
catboost==1.2.7
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.7
click-plugins==1.1.1
cligj==0.7.2
cloudpathlib==0.20.0
cloudpickle==3.1.0
cohere==4.37
comm==0.2.2
confection==0.1.5
contextily==0.9.2
contourpy==1.3.0
cycler==0.12.1
cymem==2.0.10
debugpy==1.8.6
decorator==5.1.1
defusedxml==0.7.1
Django==5.1.1
docopt==0.6.2
easydict==1.13
einops==0.8.1
et_xmlfile==2.0.0
executing==2.1.0
fastapi==0.116.1
fastavro==1.12.0
fastjsonschema==2.20.0
filelock==3.18.0
flatbuffers==24.3.25
folium==0.17.0
fonttools==4.54.1
fqdn==1.5.1
frozenlist==1.7.0
fsspec==2025.3.2
gast==0.6.0
GeoAlchemy2==0.18.0
geographiclib==2.0
geopandas==1.1.1
geopy==2.4.1
gitdb==4.0.12
GitPython==3.1.45
google-pasta==0.2.0
graphviz==0.20.3
grpcio==1.66.1
h11==0.16.0
h5py==3.11.0
hf-xet==1.2.0
holidays==0.61
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.1.7
idna==3.10
imageio==2.37.0
imbalanced-learn==0.12.4
importlib-metadata==6.11.0
iniconfig==2.1.0
ipykernel==6.29.5
ipython==8.12.3
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.7
isoduration==20.11.0
jedi==0.19.1
Jinja2==3.1.6
joblib==1.4.2
json5==0.12.1
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.12.0
jupyter-lsp==2.2.6
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.16.0
jupyter_server_terminals==0.5.3
jupyterlab==4.4.6
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.15
kagglehub==0.3.6
keras==3.5.0
keras-models==0.0.7
kiwisolver==1.4.7
langcodes==3.5.0
language_data==1.3.0
lazy_loader==0.4
libclang==18.1.1
librosa==0.11.0
lightgbm==4.5.0
lime==0.2.0.1
llvmlite==0.43.0
mapclassify==2.10.0
marisa-trie==1.2.1
Markdown==3.5.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.2
matplotlib-inline==0.1.7
mdurl==0.1.2
mercantile==1.2.1
mistune==3.1.3
ml-dtypes==0.4.1
mpmath==1.3.0
msgpack==1.1.0
multidict==6.6.4
murmurhash==1.0.11
namex==0.0.8
narwhals==2.0.1
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.5
notebook==7.4.5
notebook_shim==0.2.4
notion-client==2.2.1
npm==0.1.1
numba==0.60.0
numpy==1.26.4
opencv-python==4.10.0.84
openpyxl==3.1.5
opt-einsum==3.3.0
optional-django==0.1.0
optree==0.12.1
overrides==7.7.0
packaging==25.0
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
pathlib==1.0.1
patsy==1.0.1
pexpect==4.9.0
pickleshare==0.7.5
pillow==11.3.0
pipreqs==0.5.0
platformdirs==4.3.6
plotly==5.24.1
pluggy==1.6.0
pooch==1.8.2
preshed==3.0.9
prometheus_client==0.22.1
prompt_toolkit==3.0.48
propcache==0.4.0
protobuf==4.25.5
psutil==6.0.0
psycopg==3.2.9
psycopg-binary==3.2.9
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pyarrow==21.0.0
pycparser==2.22
pydantic==2.10.3
pydantic_core==2.27.1
pydeck==0.9.1
pydot==3.0.2
pydub==0.25.1
Pygments==2.18.0
pyogrio==0.9.0
pyparsing==3.1.4
pyproj==3.6.1
pyshp==2.3.1
pytest==7.4.3
python-bidi==0.6.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-json-logger==3.3.0
python-multipart==0.0.20
pytz==2024.2
PyYAML==6.0.2
pyzmq==26.2.0
RapidFuzz==3.13.0
rasterio==1.4.0
referencing==0.35.1
regex==2025.9.18
requests==2.31.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.8.1
rpds-py==0.20.0
safetensors==0.7.0
scikit-image==0.25.2
scikit-learn==1.5.2
scipy==1.16.1
seaborn==0.13.2
Send2Trash==1.8.3
shap==0.46.0
shapely==2.0.6
shellingham==1.5.4
six==1.16.0
slicer==0.0.8
smart-open==7.0.5
smmap==5.0.2
sniffio==1.3.1
sounddevice==0.5.2
soundfile==0.13.1
soupsieve==2.7
soxr==0.5.0.post1
spacy==3.8.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.43
sqlparse==0.5.1
srsly==2.4.8
stack-data==0.6.3
starlette==0.47.3
statsmodels==0.14.4
streamlit==1.48.0
streamlit-folium==0.25.1
sympy==1.14.0
tenacity==9.0.0
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
tensorflow-io-gcs-filesystem==0.37.1
termcolor==2.4.0
terminado==0.18.1
thinc==8.3.2
threadpoolctl==3.5.0
tifffile==2025.6.11
tiktoken==0.5.2
timm==1.0.22
tinycss2==1.4.0
toml==0.10.2
torch==2.7.0
torchaudio==2.7.0
torchvision==0.22.0
tornado==6.4.1
tqdm==4.67.0
traitlets==5.14.3
typer==0.15.1
typer-slim==0.20.0
types-python-dateutil==2.9.0.20250809
typing_extensions==4.13.2
tzdata==2024.2
ultralytics==8.3.34
ultralytics-thop==2.0.12
uri-template==1.3.0
urllib3==2.4.0
uvicorn==0.38.0
variables==0.0.1
wasabi==1.1.3
wcwidth==0.2.13
weasel==0.4.1
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.0.4
widgetsnbextension==4.0.14
wrapt==1.16.0
xyzservices==2024.9.0
yarg==0.1.9
yarl==1.20.1
zipp==3.23.0
